requirements.txt
mypy==1.10.0
mypy-extensions==1.0.0
pillow==10.3.0
pypng==0.20220715.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
qrcode==7.4.2
tomli==2.0.1
typing_extensions==4.11.0
